+++ src/engine/Actor.ts
@@ -51,7 +51,7 @@ module ex {
    * 
    * ## Actor Lifecycle
    * 
-   * An [[Actor|actor]] has a basic lifecycle that dictacts how it is initialized, updated, and drawn. Once an actor is part of a 
+   * An [[Actor|actor]] has a basic lifecycle that dictates how it is initialized, updated, and drawn. Once an actor is part of a 
    * [[Scene|scene]], it will follow this lifecycle.
    * 
    * ![Actor Lifecycle](/assets/images/docs/ActorLifecycle.png)
@@ -473,14 +473,14 @@ module ex {
     }
     
     /**
-     * Gets the current momemnt of inertia, moi can be thought of as the resistance to rotation.
+     * Gets the current moment of inertia, moi can be thought of as the resistance to rotation.
      */
     public get moi() {
        return this.body.moi;
     }
     
     /**
-     * Sets the current momemnt of inertia, moi can be thought of as the resistance to rotation.
+     * Sets the current moment of inertia, moi can be thought of as the resistance to rotation.
      */
     public set moi(theMoi: number) {
        this.body.moi = theMoi;
@@ -627,7 +627,7 @@ module ex {
     public color: Color;
 
     /**
-     * Whether or not to enable the [[CapturePointer]] trait that propogates 
+     * Whether or not to enable the [[CapturePointer]] trait that propagates 
      * pointer events to this actor
      */
     public enableCapturePointer: boolean = false;
@@ -689,14 +689,15 @@ module ex {
     }
 
     private _checkForPointerOptIn(eventName: string) {
-       if (eventName && (eventName.toLowerCase() === 'pointerdown' ||
-          eventName.toLowerCase() === 'pointerup' ||
-          eventName.toLowerCase() === 'pointermove')) {
+      if (eventName) {
+        const normalized = eventName.toLowerCase();
+        if (normalized === 'pointerup' || normalized === 'pointerdown' || normalized === 'pointermove') {
           this.enableCapturePointer = true;
-          if (eventName.toLowerCase() === 'pointermove') {
-             this.capturePointer.captureMoveEvents = true;
+          if (normalized === 'pointermove') {
+            this.capturePointer.captureMoveEvents = true;
           }
-       }
+        }
+      }
     }
    
     public on(eventName: ex.Events.kill, handler: (event?: KillEvent) => void);
@@ -791,7 +792,7 @@ module ex {
     }
      
     /**
-     * Add minimum translation vectors accumulated during the current frame to resolve collisons.
+     * Add minimum translation vectors accumulated during the current frame to resolve collisions.
      */ 
     public addMtv(mtv: Vector) {
         this._totalMtv.addEqual(mtv);
@@ -857,7 +858,7 @@ module ex {
      * Sets the z-index of an actor and updates it in the drawing list for the scene. 
      * The z-index determines the relative order an actor is drawn in.
      * Actors with a higher z-index are drawn on top of actors with a lower z-index
-     * @param actor The child actor to remove
+     * @param newIndex new z-index to assign
      */
      public setZIndex(newIndex: number) {
        this.scene.cleanupDrawTree(this);
@@ -1173,10 +1174,8 @@ module ex {
           !(this instanceof Label)) {
           totalAcc.addEqual(ex.Physics.acc);
        }
-
-       this.oldVel = this.vel;
-       this.vel.addEqual(totalAcc.scale(seconds));
-
+       
+       this.vel.addEqual(totalAcc.scale(seconds));       
        this.pos.addEqual(this.vel.scale(seconds)).addEqual(totalAcc.scale(0.5 * seconds * seconds));
 
        this.rx += this.torque * (1.0 / this.moi) * seconds;
@@ -1229,6 +1228,11 @@ module ex {
           }
        }
 
+       // Capture old values before integration step updates them
+       this.oldVel.setTo(this.vel.x, this.vel.y);
+       this.oldPos.setTo(this.pos.x, this.pos.y);
+
+       // Run Euler integration
        this.integrate(delta);
 
        // Update actor pipeline (movement, collision detection, event propagation, offscreen culling)