+++ src/engine/Engine.ts
@@ -230,8 +230,8 @@ module ex {
     *     |_ Actor 1
     *       |_ Child Actor 1
     *     |_ Actor 2
-    *   |_ Scene 2 (deactiveated)
-    *   |_ Scene 3 (deactiveated)
+    *   |_ Scene 2 (deactivated)
+    *   |_ Scene 3 (deactivated)
     * ```
     *
     * The engine splits the game into two primary responsibilities: updating and drawing. This is
@@ -930,7 +930,7 @@ O|===|* >________________>\n\
             screenY -= focus.y;
          }
 
-         // transfrom back on zoom
+         // transform back on zoom
          screenX = screenX + this.getWidth() / 2;
          screenY = screenY + this.getHeight() / 2;
 
@@ -1068,7 +1068,7 @@ O|===|* >________________>\n\
 
       /**
        * Draws the entire game
-       * @param draw  Number of milliseconds elapsed since the last draw.
+       * @param delta  Number of milliseconds elapsed since the last draw.
        */
       private _draw(delta: number) {
          var ctx = this.ctx;