+++ src/engine/Group.ts
@@ -13,7 +13,7 @@ module ex {
     *
     * ## Using Groups
     *
-    * Groups can be used to detect collisions across a large nubmer of actors. For example 
+    * Groups can be used to detect collisions across a large number of actors. For example 
     * perhaps a large group of "enemy" actors.
     *
     * ```typescript